Clean up process source code
